Rust: Use unsigned 16 bit integers for timer IDs

+++ src/sccd/cd/codegen/sccdlib.rs
@@ -3,7 +3,7 @@ use std::cmp::Ordering;
 
 type Timestamp = usize;
 
-type TimerId = usize;
+type TimerId = u16;
 
 pub trait State<TimersType, Handle> {
   // Execute enter actions of only this state
